+++ $notes/changelog.md
@@ -18,6 +18,7 @@
 - Diet classification fixes for pudding and strawberries
 - Coarse dirt now duplicable with tickets
 - Exotic bird eggs are now useable for more egg recipes
+- Minecraft flowers now grow faster in botany pots and many modded flowers are now growable in botany pots
 # 4.1.8:
 ## Mod Added:
 - effortlessbuilding